chore: fixing tests
