+++ packages/unhead/src/plugins/sort.ts
@@ -1,54 +1,8 @@
 import { defineHeadPlugin, SortModifiers, tagWeight } from '@unhead/shared'
 
-const importRe = /@import/
-const isTruthy = (val?: string | boolean) => val === '' || val === true
 
-export default defineHeadPlugin(head => ({
+export default defineHeadPlugin((head => ({
   hooks: {
-    // capo sorting
-    'tags:beforeResolve': ({ tags }) => {
-      if (!head.ssr || head.resolvedOptions.disableCapoSorting) {
-        return
-      }
-      for (const tag of tags) {
-        if (tag.tagPosition && tag.tagPosition !== 'head')
-          continue
-        tag.tagPriority = tag.tagPriority || tagWeight(tag)
-        // skip if already prioritised
-        if (tag.tagPriority !== 100)
-          continue
-
-        const isScript = tag.tag === 'script'
-        const isLink = tag.tag === 'link'
-        if (isScript && isTruthy(tag.props.async)) {
-          // ASYNC_SCRIPT
-          tag.tagPriority = 30
-          // SYNC_SCRIPT
-        }
-        else if (tag.tag === 'style' && tag.innerHTML && importRe.test(tag.innerHTML)) {
-          // IMPORTED_STYLES
-          tag.tagPriority = 40
-        }
-        else if (isScript && tag.props.src && !isTruthy(tag.props.defer) && !isTruthy(tag.props.async) && tag.props.type !== 'module' && !tag.props.type?.endsWith('json')) {
-          tag.tagPriority = 50
-        }
-        else if ((isLink && tag.props.rel === 'stylesheet') || tag.tag === 'style') {
-          // SYNC_STYLES
-          tag.tagPriority = 60
-        }
-        else if (isLink && (tag.props.rel === 'preload' || tag.props.rel === 'modulepreload')) {
-          // PRELOAD
-          tag.tagPriority = 70
-        }
-        else if (isScript && isTruthy(tag.props.defer) && tag.props.src && !isTruthy(tag.props.async)) {
-          // DEFER_SCRIPT
-          tag.tagPriority = 80
-        }
-        else if (isLink && (tag.props.rel === 'prefetch' || tag.props.rel === 'dns-prefetch' || tag.props.rel === 'prerender')) {
-          tag.tagPriority = 90
-        }
-      }
-    },
     'tags:resolve': (ctx) => {
       // 2a. Sort based on priority
       // now we need to check render priority for each before: rule and use the dedupe key index
@@ -78,8 +32,8 @@ export default defineHeadPlugin(head => ({
       }
 
       ctx.tags.sort((a, b) => {
-        const aWeight = tagWeight(a)
-        const bWeight = tagWeight(b)
+        const aWeight = tagWeight(head, a)
+        const bWeight = tagWeight(head, b)
 
         // 2c. sort based on critical tags
         if (aWeight < bWeight) {
@@ -94,4 +48,4 @@ export default defineHeadPlugin(head => ({
       })
     },
   },
-}))
+})))
