+++ packages/vue/test/e2e/basic.test.ts
@@ -90,13 +90,13 @@ describe('vue e2e', () => {
         "headTags": "<meta charset="utf-8">
       <title>Home</title>
       <script src="https://analytics.example.com/script.js" defer async></script>
+      <script src="https://my-app.com/home.js"></script>
       <meta property="og:title" content="My amazing site">
       <meta property="og:description" content="This is my amazing site">
       <meta property="og:locale" content="en_US">
       <meta property="og:locale" content="en_AU">
       <meta property="og:image" content="https://cdn.example.com/image.jpg">
       <meta property="og:image" content="https://cdn.example.com/image2.jpg">
-      <script src="https://my-app.com/home.js"></script>
       <meta name="description" content="This is the home page">
       <script id="unhead:payload" type="application/json">{"title":"My amazing site"}</script>",
         "htmlAttrs": " lang="en"",
@@ -131,13 +131,13 @@ describe('vue e2e', () => {
       <meta charset="utf-8">
       <title>Home</title>
       <script src="https://analytics.example.com/script.js" defer="" async=""></script>
+      <script src="https://my-app.com/home.js"></script>
       <meta property="og:title" content="Home">
       <meta property="og:description" content="This is my amazing site">
       <meta property="og:locale" content="en_US">
       <meta property="og:locale" content="en_AU">
       <meta property="og:image" content="https://cdn.example.com/image.jpg">
       <meta property="og:image" content="https://cdn.example.com/image2.jpg">
-      <script src="https://my-app.com/home.js"></script>
       <meta name="description" content="This is the home page">
       <script id="unhead:payload" type="application/json">{"title":"My amazing site"}</script>
       </head>
