+++ test/unhead/resolveTags.test.ts
@@ -58,6 +58,14 @@ describe('resolveTags', () => {
           "tag": "meta",
         },
         {
+          "_e": 0,
+          "_p": 2,
+          "props": {
+            "src": "https://cdn.example.com/script.js",
+          },
+          "tag": "script",
+        },
+        {
           "_d": "htmlAttrs",
           "_e": 0,
           "_p": 0,
@@ -75,14 +83,6 @@ describe('resolveTags', () => {
             "class": "dark",
           },
           "tag": "bodyAttrs",
-        },
-        {
-          "_e": 0,
-          "_p": 2,
-          "props": {
-            "src": "https://cdn.example.com/script.js",
-          },
-          "tag": "script",
         },
         {
           "_e": 0,
@@ -96,56 +96,6 @@ describe('resolveTags', () => {
         },
       ]
     `)
-    expect(tags).toMatchInlineSnapshot(`
-      [
-        {
-          "_d": "charset",
-          "_e": 0,
-          "_p": 3,
-          "props": {
-            "charset": "utf-8",
-          },
-          "tag": "meta",
-        },
-        {
-          "_d": "htmlAttrs",
-          "_e": 0,
-          "_p": 0,
-          "props": {
-            "dir": "ltr",
-            "lang": "en",
-          },
-          "tag": "htmlAttrs",
-        },
-        {
-          "_d": "bodyAttrs",
-          "_e": 0,
-          "_p": 1,
-          "props": {
-            "class": "dark",
-          },
-          "tag": "bodyAttrs",
-        },
-        {
-          "_e": 0,
-          "_p": 2,
-          "props": {
-            "src": "https://cdn.example.com/script.js",
-          },
-          "tag": "script",
-        },
-        {
-          "_e": 0,
-          "_p": 4,
-          "props": {
-            "href": "https://cdn.example.com/favicon.ico",
-            "rel": "icon",
-            "type": "image/x-icon",
-          },
-          "tag": "link",
-        },
-      ]
-    `, 'old')
   })
 
   it('basic /w removal', async () => {
@@ -207,6 +157,14 @@ describe('resolveTags', () => {
           "tag": "meta",
         },
         {
+          "_e": 0,
+          "_p": 2,
+          "props": {
+            "src": "https://cdn.example.com/script2.js",
+          },
+          "tag": "script",
+        },
+        {
           "_d": "htmlAttrs",
           "_e": 0,
           "_p": 0,
@@ -224,14 +182,6 @@ describe('resolveTags', () => {
             "class": "dark",
           },
           "tag": "bodyAttrs",
-        },
-        {
-          "_e": 0,
-          "_p": 2,
-          "props": {
-            "src": "https://cdn.example.com/script2.js",
-          },
-          "tag": "script",
         },
         {
           "_e": 0,
