+++ test/unhead/ssr/tagPriority.test.ts
@@ -38,7 +38,6 @@ describe('tag priority', () => {
             "src": "/not-important-script.js",
           },
           "tag": "script",
-          "tagPriority": 50,
         },
       ]
     `)
